split kernels and dispatchers
