+++ torchvision/prototype/transforms/__init__.py
@@ -1,4 +1,4 @@
-from . import functional
-from .functional import InterpolationMode  # usort: skip
-
+from . import kernels  # usort: skip
+from . import functional  # usort: skip
 from ._presets import CocoEval, ImageNetEval, VocEval, Kinect400Eval, RaftEval
+from .functional import InterpolationMode
