+++ torchvision/prototype/transforms/functional/_geometry.py
@@ -1,35 +1,22 @@
-from typing import Tuple, List, Optional, TypeVar, Any
+from typing import List, Optional, TypeVar, Any
 
 import torch
 from torchvision.prototype import features
+from torchvision.prototype.transforms import kernels as K
 from torchvision.transforms import functional as _F, InterpolationMode
 
-from ._meta_conversion import convert_bounding_box_format
 from ._utils import dispatch, FEATURE_SPECIFIC_DEFAULT
 
 T = TypeVar("T", bound=features.Feature)
 
 
-horizontal_flip_image = _F.hflip
-
-
-def horizontal_flip_bounding_box(
-    bounding_box: torch.Tensor, *, format: features.BoundingBoxFormat, image_size: Tuple[int, int]
-) -> torch.Tensor:
-    bounding_box = convert_bounding_box_format(
-        bounding_box, old_format=format, new_format=features.BoundingBoxFormat.XYXY, copy=True
-    )
-    bounding_box[..., (0, 2)] = image_size[1] - bounding_box[..., (2, 0)]
-    return convert_bounding_box_format(bounding_box, old_format=features.BoundingBoxFormat.XYXY, new_format=format)
-
-
 def _horizontal_flip_bounding_box(input: features.BoundingBox) -> torch.Tensor:
-    return horizontal_flip_bounding_box(input, format=input.format, image_size=input.image_size)
+    return K.horizontal_flip_bounding_box(input, format=input.format, image_size=input.image_size)
 
 
 @dispatch(
     {
-        features.Image: horizontal_flip_image,
+        features.Image: K.horizontal_flip_image,
         features.BoundingBox: _horizontal_flip_bounding_box,
     },
     pil_kernel=_F.hflip,
@@ -39,56 +26,15 @@ def horizontal_flip(input: T) -> T:
     pass
 
 
-def resize_image(
-    image: torch.Tensor,
-    size: List[int],
-    interpolation: InterpolationMode = InterpolationMode.BILINEAR,
-    max_size: Optional[int] = None,
-    antialias: Optional[bool] = None,
-) -> torch.Tensor:
-    new_height, new_width = size
-    num_channels, old_height, old_width = image.shape[-3:]
-    batch_shape = image.shape[:-3]
-    return _F.resize(
-        image.reshape((-1, num_channels, old_height, old_width)),
-        size=size,
-        interpolation=interpolation,
-        max_size=max_size,
-        antialias=antialias,
-    ).reshape(batch_shape + (num_channels, new_height, new_width))
-
-
-def resize_segmentation_mask(
-    segmentation_mask: torch.Tensor,
-    size: List[int],
-    interpolation: InterpolationMode = InterpolationMode.NEAREST,
-    max_size: Optional[int] = None,
-    antialias: Optional[bool] = None,
-) -> torch.Tensor:
-    return resize_image(
-        segmentation_mask, size=size, interpolation=interpolation, max_size=max_size, antialias=antialias
-    )
-
-
-# TODO: handle max_size
-def resize_bounding_box(
-    bounding_box: torch.Tensor, *, old_image_size: List[int], new_image_size: List[int]
-) -> torch.Tensor:
-    old_height, old_width = old_image_size
-    new_height, new_width = new_image_size
-    ratios = torch.tensor((new_width / old_width, new_height / old_height))
-    return bounding_box.view(-1, 2, 2).mul(ratios).view(bounding_box.shape)
-
-
 def _resize_bounding_box(input: features.BoundingBox, *, size: List[int], **_: Any) -> features.BoundingBox:
-    output = resize_bounding_box(input, old_image_size=list(input.image_size), new_image_size=size)
+    output = K.resize_bounding_box(input, old_image_size=list(input.image_size), new_image_size=size)
     return features.BoundingBox.new_like(input, output, image_size=size)
 
 
 @dispatch(
     {
-        features.Image: resize_image,
-        features.SegmentationMask: resize_segmentation_mask,
+        features.Image: K.resize_image,
+        features.SegmentationMask: K.resize_segmentation_mask,
         features.BoundingBox: _resize_bounding_box,
     },
     pil_kernel=_F.resize,
@@ -105,12 +51,9 @@ def resize(
     pass
 
 
-center_crop_image = _F.center_crop
-
-
 @dispatch(
     {
-        features.Image: center_crop_image,
+        features.Image: K.center_crop_image,
     }
 )
 def center_crop(input: T, *, output_size: List[int]) -> T:
@@ -118,12 +61,9 @@ def center_crop(input: T, *, output_size: List[int]) -> T:
     pass
 
 
-resized_crop_image = _F.resized_crop
-
-
 @dispatch(
     {
-        features.Image: resized_crop_image,
+        features.Image: K.resized_crop_image,
     }
 )
 def resized_crop(
@@ -140,12 +80,9 @@ def resized_crop(
     pass
 
 
-affine_image = _F.affine
-
-
 @dispatch(
     {
-        features.Image: affine_image,
+        features.Image: K.affine_image,
     }
 )
 def affine(
@@ -165,12 +102,9 @@ def affine(
     pass
 
 
-rotate_image = _F.rotate
-
-
 @dispatch(
     {
-        features.Image: rotate_image,
+        features.Image: K.rotate_image,
     }
 )
 def rotate(
