+++ torchvision/prototype/transforms/kernels/__init__.py
@@ -1,0 +1,33 @@
+from ._meta_conversion import convert_bounding_box_format, convert_color_space  # usort: skip
+
+from ._augment import (
+    erase_image,
+    mixup_image,
+    mixup_one_hot_label,
+    cutmix_image,
+    cutmix_one_hot_label,
+)
+from ._color import (
+    adjust_brightness_image,
+    adjust_contrast_image,
+    adjust_saturation_image,
+    adjust_sharpness_image,
+    posterize_image,
+    solarize_image,
+    autocontrast_image,
+    equalize_image,
+    invert_image,
+)
+from ._geometry import (
+    horizontal_flip_bounding_box,
+    horizontal_flip_image,
+    resize_bounding_box,
+    resize_image,
+    resize_segmentation_mask,
+    center_crop_image,
+    resized_crop_image,
+    affine_image,
+    rotate_image,
+)
+from ._misc import normalize_image
+from ._type_conversion import decode_image_with_pil, decode_video_with_av, label_to_one_hot
