+++ torchvision/prototype/transforms/kernels/_color.py
@@ -1,0 +1,12 @@
+from torchvision.transforms import functional as _F
+
+
+adjust_brightness_image = _F.adjust_brightness
+adjust_saturation_image = _F.adjust_saturation
+adjust_contrast_image = _F.adjust_contrast
+adjust_sharpness_image = _F.adjust_sharpness
+posterize_image = _F.posterize
+solarize_image = _F.solarize
+autocontrast_image = _F.autocontrast
+equalize_image = _F.equalize
+invert_image = _F.invert
