+++ torchvision/prototype/transforms/kernels/_misc.py
@@ -2,16 +2,19 @@ from typing import List
 from typing import TypeVar
 
 from torchvision.prototype import features
-from torchvision.prototype.transforms import kernels as K
+from torchvision.transforms import functional as _F
 
 from ._utils import dispatch
 
 T = TypeVar("T", bound=features.Feature)
 
 
+normalize_image = _F.normalize
+
+
 @dispatch(
     {
-        features.Image: K.normalize_image,
+        features.Image: normalize_image,
     }
 )
 def normalize(input: T, *, mean: List[float], std: List[float], inplace: bool = False) -> T:
